Refactored prisoner event data handling and introduced EventResult.
Moved `PrisonerEventData` and `PrisonerResponseEntry` to a new `records` package for better organization. Introduced the `EventResult` record to encapsulate event response details, replacing the use of `EventResultEffect` in data models. Updated YAML files and related classes to align with this refactor.
