+++ MekHQ/src/mekhq/campaign/personnel/randomEvents/RandomEventLibraries.java
@@ -2,7 +2,7 @@ package mekhq.campaign.personnel.randomEvents;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import mekhq.campaign.personnel.prisoners.PrisonerEventData;
+import mekhq.campaign.personnel.prisoners.records.PrisonerEventData;
 import mekhq.campaign.personnel.prisoners.yaml.PrisonerEventDataWrapper;
 
 import java.io.File;
